+++ x/emissions/keeper/keeper.go
@@ -1292,6 +1292,12 @@ func (k *Keeper) AppendInference(
 		return types.ErrCantUpdateEmaMoreThanOncePerWindow
 	}
 
+	// Penalise the inferer if needed
+	previousEmaScore, err = k.ApplyLivenessPenaltyToInferer(ctx, topic, nonceBlockHeight, previousEmaScore)
+	if err != nil {
+		return errorsmod.Wrap(err, "error trying to penalise inferer")
+	}
+
 	// Check if the inferer is new and set initial EMA score
 	if previousEmaScore.BlockHeight == 0 {
 		initialEmaScore, err := k.GetTopicInitialInfererEmaScore(ctx, topic.Id)
@@ -1456,6 +1462,12 @@ func (k *Keeper) AppendForecast(
 	// Only calc and save if there's a new update
 	if previousEmaScore.BlockHeight >= nonceBlockHeight {
 		return types.ErrCantUpdateEmaMoreThanOncePerWindow
+	}
+
+	// Penalise the forecaster if needed
+	previousEmaScore, err = k.ApplyLivenessPenaltyToForecaster(ctx, topic, nonceBlockHeight, previousEmaScore)
+	if err != nil {
+		return errorsmod.Wrap(err, "error trying to penalise forecaster")
 	}
 
 	// Check if the forecaster is new and set initial EMA score
@@ -1669,6 +1681,12 @@ func (k *Keeper) AppendReputerLoss(
 	// Only calc and save if there's a new update
 	if previousEmaScore.BlockHeight >= nonceBlockHeight {
 		return types.ErrCantUpdateEmaMoreThanOncePerWindow
+	}
+
+	// Penalise the reputer if needed
+	previousEmaScore, err = k.ApplyLivenessPenaltyToReputer(ctx, topic, nonceBlockHeight, previousEmaScore)
+	if err != nil {
+		return errorsmod.Wrap(err, "error trying to penalise reputer")
 	}
 
 	// Check if the reputer is new and set initial EMA score